% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_application}
\alias{applicationinsights_describe_application}
\title{Describes the application}
\usage{
applicationinsights_describe_application(ResourceGroupName)
}
\arguments{
\item{ResourceGroupName}{[required] The name of the resource group.}
}
\description{
Describes the application.

See \url{https://paws-r.github.io/docs/applicationinsights/describe_application.html} for full documentation.
}
\keyword{internal}
