% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_describe_observation}
\alias{applicationinsights_describe_observation}
\title{Describes an anomaly or error with the application}
\usage{
applicationinsights_describe_observation(ObservationId)
}
\arguments{
\item{ObservationId}{[required] The ID of the observation.}
}
\description{
Describes an anomaly or error with the application.

See \url{https://paws-r.github.io/docs/applicationinsights/describe_observation.html} for full documentation.
}
\keyword{internal}
