% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_list_applications}
\alias{applicationinsights_list_applications}
\title{Lists the IDs of the applications that you are monitoring}
\usage{
applicationinsights_list_applications(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}
}
\description{
Lists the IDs of the applications that you are monitoring.

See \url{https://paws-r.github.io/docs/applicationinsights/list_applications.html} for full documentation.
}
\keyword{internal}
