% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applicationinsights_operations.R
\name{applicationinsights_list_problems}
\alias{applicationinsights_list_problems}
\title{Lists the problems with your application}
\usage{
applicationinsights_list_problems(
  ResourceGroupName = NULL,
  StartTime = NULL,
  EndTime = NULL,
  MaxResults = NULL,
  NextToken = NULL,
  ComponentName = NULL
)
}
\arguments{
\item{ResourceGroupName}{The name of the resource group.}

\item{StartTime}{The time when the problem was detected, in epoch seconds. If you don't
specify a time frame for the request, problems within the past seven
days are returned.}

\item{EndTime}{The time when the problem ended, in epoch seconds. If not specified,
problems within the past seven days are returned.}

\item{MaxResults}{The maximum number of results to return in a single call. To retrieve
the remaining results, make another call with the returned \code{NextToken}
value.}

\item{NextToken}{The token to request the next page of results.}

\item{ComponentName}{The name of the component.}
}
\description{
Lists the problems with your application.

See \url{https://paws-r.github.io/docs/applicationinsights/list_problems.html} for full documentation.
}
\keyword{internal}
