% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_batch_create_delegation_by_assessment}
\alias{auditmanager_batch_create_delegation_by_assessment}
\title{Creates a batch of delegations for an assessment in Audit Manager}
\usage{
auditmanager_batch_create_delegation_by_assessment(
  createDelegationRequests,
  assessmentId
)
}
\arguments{
\item{createDelegationRequests}{[required] The API request to batch create delegations in Audit Manager.}

\item{assessmentId}{[required] The identifier for the assessment.}
}
\description{
Creates a batch of delegations for an assessment in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/batch_create_delegation_by_assessment.html} for full documentation.
}
\keyword{internal}
