% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_create_assessment_report}
\alias{auditmanager_create_assessment_report}
\title{Creates an assessment report for the specified assessment}
\usage{
auditmanager_create_assessment_report(name, description = NULL, assessmentId)
}
\arguments{
\item{name}{[required] The name of the new assessment report.}

\item{description}{The description of the assessment report.}

\item{assessmentId}{[required] The identifier for the assessment.}
}
\description{
Creates an assessment report for the specified assessment.

See \url{https://paws-r.github.io/docs/auditmanager/create_assessment_report.html} for full documentation.
}
\keyword{internal}
