% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_assessment_report}
\alias{auditmanager_delete_assessment_report}
\title{Deletes an assessment report in Audit Manager}
\usage{
auditmanager_delete_assessment_report(assessmentId, assessmentReportId)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the assessment.}

\item{assessmentReportId}{[required] The unique identifier for the assessment report.}
}
\description{
Deletes an assessment report in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/delete_assessment_report.html} for full documentation.
}
\keyword{internal}
