% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_delete_control}
\alias{auditmanager_delete_control}
\title{Deletes a custom control in Audit Manager}
\usage{
auditmanager_delete_control(controlId)
}
\arguments{
\item{controlId}{[required] The unique identifier for the control.}
}
\description{
Deletes a custom control in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/delete_control.html} for full documentation.
}
\keyword{internal}
