% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_deregister_organization_admin_account}
\alias{auditmanager_deregister_organization_admin_account}
\title{Removes the specified Amazon Web Services account as a delegated
administrator for Audit Manager}
\usage{
auditmanager_deregister_organization_admin_account(adminAccountId = NULL)
}
\arguments{
\item{adminAccountId}{The identifier for the administrator account.}
}
\description{
Removes the specified Amazon Web Services account as a delegated administrator for Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/deregister_organization_admin_account.html} for full documentation.
}
\keyword{internal}
