% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_assessment_report_url}
\alias{auditmanager_get_assessment_report_url}
\title{Returns the URL of an assessment report in Audit Manager}
\usage{
auditmanager_get_assessment_report_url(assessmentReportId, assessmentId)
}
\arguments{
\item{assessmentReportId}{[required] The unique identifier for the assessment report.}

\item{assessmentId}{[required] The unique identifier for the assessment.}
}
\description{
Returns the URL of an assessment report in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/get_assessment_report_url.html} for full documentation.
}
\keyword{internal}
