% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_get_control}
\alias{auditmanager_get_control}
\title{Returns a control from Audit Manager}
\usage{
auditmanager_get_control(controlId)
}
\arguments{
\item{controlId}{[required] The identifier for the control.}
}
\description{
Returns a control from Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/get_control.html} for full documentation.
}
\keyword{internal}
