% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_assessment_control_insights_by_control_domain}
\alias{auditmanager_list_assessment_control_insights_by_control_domain}
\title{Lists the latest analytics data for controls within a specific control
domain and a specific active assessment}
\usage{
auditmanager_list_assessment_control_insights_by_control_domain(
  controlDomainId,
  assessmentId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{controlDomainId}{[required] The unique identifier for the control domain.}

\item{assessmentId}{[required] The unique identifier for the active assessment.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Lists the latest analytics data for controls within a specific control domain and a specific active assessment.

See \url{https://paws-r.github.io/docs/auditmanager/list_assessment_control_insights_by_control_domain.html} for full documentation.
}
\keyword{internal}
