% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_control_domain_insights_by_assessment}
\alias{auditmanager_list_control_domain_insights_by_assessment}
\title{Lists analytics data for control domains within a specified active
assessment}
\usage{
auditmanager_list_control_domain_insights_by_assessment(
  assessmentId,
  nextToken = NULL,
  maxResults = NULL
)
}
\arguments{
\item{assessmentId}{[required] The unique identifier for the active assessment.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Lists analytics data for control domains within a specified active assessment.

See \url{https://paws-r.github.io/docs/auditmanager/list_control_domain_insights_by_assessment.html} for full documentation.
}
\keyword{internal}
