% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_list_controls}
\alias{auditmanager_list_controls}
\title{Returns a list of controls from Audit Manager}
\usage{
auditmanager_list_controls(controlType, nextToken = NULL, maxResults = NULL)
}
\arguments{
\item{controlType}{[required] The type of control, such as a standard control or a custom control.}

\item{nextToken}{The pagination token that's used to fetch the next set of results.}

\item{maxResults}{Represents the maximum number of results on a page or for an API request
call.}
}
\description{
Returns a list of controls from Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/list_controls.html} for full documentation.
}
\keyword{internal}
