% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_start_assessment_framework_share}
\alias{auditmanager_start_assessment_framework_share}
\title{Creates a share request for a custom framework in Audit Manager}
\usage{
auditmanager_start_assessment_framework_share(
  frameworkId,
  destinationAccount,
  destinationRegion,
  comment = NULL
)
}
\arguments{
\item{frameworkId}{[required] The unique identifier for the custom framework to be shared.}

\item{destinationAccount}{[required] The Amazon Web Services account of the recipient.}

\item{destinationRegion}{[required] The Amazon Web Services Region of the recipient.}

\item{comment}{An optional comment from the sender about the share request.}
}
\description{
Creates a share request for a custom framework in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/start_assessment_framework_share.html} for full documentation.
}
\keyword{internal}
