% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_update_control}
\alias{auditmanager_update_control}
\title{Updates a custom control in Audit Manager}
\usage{
auditmanager_update_control(
  controlId,
  name,
  description = NULL,
  testingInformation = NULL,
  actionPlanTitle = NULL,
  actionPlanInstructions = NULL,
  controlMappingSources
)
}
\arguments{
\item{controlId}{[required] The identifier for the control.}

\item{name}{[required] The name of the updated control.}

\item{description}{The optional description of the control.}

\item{testingInformation}{The steps that you should follow to determine if the control is met.}

\item{actionPlanTitle}{The title of the action plan for remediating the control.}

\item{actionPlanInstructions}{The recommended actions to carry out if the control isn't fulfilled.}

\item{controlMappingSources}{[required] The data mapping sources for the control.}
}
\description{
Updates a custom control in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/update_control.html} for full documentation.
}
\keyword{internal}
