% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditmanager_operations.R
\name{auditmanager_validate_assessment_report_integrity}
\alias{auditmanager_validate_assessment_report_integrity}
\title{Validates the integrity of an assessment report in Audit Manager}
\usage{
auditmanager_validate_assessment_report_integrity(s3RelativePath)
}
\arguments{
\item{s3RelativePath}{[required] The relative path of the Amazon S3 bucket that the assessment report is
stored in.}
}
\description{
Validates the integrity of an assessment report in Audit Manager.

See \url{https://paws-r.github.io/docs/auditmanager/validate_assessment_report_integrity.html} for full documentation.
}
\keyword{internal}
