% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_launch_configurations}
\alias{autoscaling_describe_launch_configurations}
\title{Gets information about the launch configurations in the account and
Region}
\usage{
autoscaling_describe_launch_configurations(
  LaunchConfigurationNames = NULL,
  NextToken = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{LaunchConfigurationNames}{The launch configuration names. If you omit this property, all launch
configurations are described.

Array Members: Maximum number of 50 items.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{50} and the maximum value is \code{100}.}
}
\description{
Gets information about the launch configurations in the account and Region.

See \url{https://paws-r.github.io/docs/autoscaling/describe_launch_configurations.html} for full documentation.
}
\keyword{internal}
