% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_load_balancers}
\alias{autoscaling_describe_load_balancers}
\title{Gets information about the load balancers for the specified Auto Scaling
group}
\usage{
autoscaling_describe_load_balancers(
  AutoScalingGroupName,
  NextToken = NULL,
  MaxRecords = NULL
)
}
\arguments{
\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{100} and the maximum value is \code{100}.}
}
\description{
Gets information about the load balancers for the specified Auto Scaling group.

See \url{https://paws-r.github.io/docs/autoscaling/describe_load_balancers.html} for full documentation.
}
\keyword{internal}
