% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_scaling_activities}
\alias{autoscaling_describe_scaling_activities}
\title{Gets information about the scaling activities in the account and Region}
\usage{
autoscaling_describe_scaling_activities(
  ActivityIds = NULL,
  AutoScalingGroupName = NULL,
  IncludeDeletedGroups = NULL,
  MaxRecords = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ActivityIds}{The activity IDs of the desired scaling activities. If you omit this
property, all activities for the past six weeks are described. If
unknown activities are requested, they are ignored with no error. If you
specify an Auto Scaling group, the results are limited to that group.

Array Members: Maximum number of 50 IDs.}

\item{AutoScalingGroupName}{The name of the Auto Scaling group.}

\item{IncludeDeletedGroups}{Indicates whether to include scaling activity from deleted Auto Scaling
groups.}

\item{MaxRecords}{The maximum number of items to return with this call. The default value
is \code{100} and the maximum value is \code{100}.}

\item{NextToken}{The token for the next set of items to return. (You received this token
from a previous call.)}
}
\description{
Gets information about the scaling activities in the account and Region.

See \url{https://paws-r.github.io/docs/autoscaling/describe_scaling_activities.html} for full documentation.
}
\keyword{internal}
