% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_describe_termination_policy_types}
\alias{autoscaling_describe_termination_policy_types}
\title{Describes the termination policies supported by Amazon EC2 Auto Scaling}
\usage{
autoscaling_describe_termination_policy_types()
}
\description{
Describes the termination policies supported by Amazon EC2 Auto Scaling.

See \url{https://paws-r.github.io/docs/autoscaling/describe_termination_policy_types.html} for full documentation.
}
\keyword{internal}
