% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_enter_standby}
\alias{autoscaling_enter_standby}
\title{Moves the specified instances into the standby state}
\usage{
autoscaling_enter_standby(
  InstanceIds = NULL,
  AutoScalingGroupName,
  ShouldDecrementDesiredCapacity
)
}
\arguments{
\item{InstanceIds}{The IDs of the instances. You can specify up to 20 instances.}

\item{AutoScalingGroupName}{[required] The name of the Auto Scaling group.}

\item{ShouldDecrementDesiredCapacity}{[required] Indicates whether to decrement the desired capacity of the Auto Scaling
group by the number of instances moved to \code{Standby} mode.}
}
\description{
Moves the specified instances into the standby state.

See \url{https://paws-r.github.io/docs/autoscaling/enter_standby.html} for full documentation.
}
\keyword{internal}
