% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_operations.R
\name{autoscaling_terminate_instance_in_auto_scaling_group}
\alias{autoscaling_terminate_instance_in_auto_scaling_group}
\title{Terminates the specified instance and optionally adjusts the desired
group size}
\usage{
autoscaling_terminate_instance_in_auto_scaling_group(
  InstanceId,
  ShouldDecrementDesiredCapacity
)
}
\arguments{
\item{InstanceId}{[required] The ID of the instance.}

\item{ShouldDecrementDesiredCapacity}{[required] Indicates whether terminating the instance also decrements the size of
the Auto Scaling group.}
}
\description{
Terminates the specified instance and optionally adjusts the desired group size. This operation cannot be called on instances in a warm pool.

See \url{https://paws-r.github.io/docs/autoscaling/terminate_instance_in_auto_scaling_group.html} for full documentation.
}
\keyword{internal}
