% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_deactivate_type}
\alias{cloudformation_deactivate_type}
\title{Deactivates a public extension that was previously activated in this
account and region}
\usage{
cloudformation_deactivate_type(TypeName = NULL, Type = NULL, Arn = NULL)
}
\arguments{
\item{TypeName}{The type name of the extension, in this account and region. If you
specified a type name alias when enabling the extension, use the type
name alias.

Conditional: You must specify either \code{Arn}, or \code{TypeName} and \code{Type}.}

\item{Type}{The extension type.

Conditional: You must specify either \code{Arn}, or \code{TypeName} and \code{Type}.}

\item{Arn}{The Amazon Resource Name (ARN) for the extension, in this account and
region.

Conditional: You must specify either \code{Arn}, or \code{TypeName} and \code{Type}.}
}
\description{
Deactivates a public extension that was previously activated in this account and region.

See \url{https://paws-r.github.io/docs/cloudformation/deactivate_type.html} for full documentation.
}
\keyword{internal}
