% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudformation_operations.R
\name{cloudformation_deregister_type}
\alias{cloudformation_deregister_type}
\title{Marks an extension or extension version as DEPRECATED in the
CloudFormation registry, removing it from active use}
\usage{
cloudformation_deregister_type(
  Arn = NULL,
  Type = NULL,
  TypeName = NULL,
  VersionId = NULL
)
}
\arguments{
\item{Arn}{The Amazon Resource Name (ARN) of the extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{Type}{The kind of extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{TypeName}{The name of the extension.

Conditional: You must specify either \code{TypeName} and \code{Type}, or \code{Arn}.}

\item{VersionId}{The ID of a specific version of the extension. The version ID is the
value at the end of the Amazon Resource Name (ARN) assigned to the
extension version when it is registered.}
}
\description{
Marks an extension or extension version as \code{DEPRECATED} in the CloudFormation registry, removing it from active use. Deprecated extensions or extension versions cannot be used in CloudFormation operations.

See \url{https://paws-r.github.io/docs/cloudformation/deregister_type.html} for full documentation.
}
\keyword{internal}
