% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_deauthorize_connection}
\alias{cloudwatchevents_deauthorize_connection}
\title{Removes all authorization parameters from the connection}
\usage{
cloudwatchevents_deauthorize_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to remove authorization from.}
}
\description{
Removes all authorization parameters from the connection. This lets you remove the secret from the connection so you can reuse it without having to create a new connection.

See \url{https://paws-r.github.io/docs/cloudwatchevents/deauthorize_connection.html} for full documentation.
}
\keyword{internal}
