% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_describe_connection}
\alias{cloudwatchevents_describe_connection}
\title{Retrieves details about a connection}
\usage{
cloudwatchevents_describe_connection(Name)
}
\arguments{
\item{Name}{[required] The name of the connection to retrieve.}
}
\description{
Retrieves details about a connection.

See \url{https://paws-r.github.io/docs/cloudwatchevents/describe_connection.html} for full documentation.
}
\keyword{internal}
