% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_describe_rule}
\alias{cloudwatchevents_describe_rule}
\title{Describes the specified rule}
\usage{
cloudwatchevents_describe_rule(Name, EventBusName = NULL)
}
\arguments{
\item{Name}{[required] The name of the rule.}

\item{EventBusName}{The name or ARN of the event bus associated with the rule. If you omit
this, the default event bus is used.}
}
\description{
Describes the specified rule.

See \url{https://paws-r.github.io/docs/cloudwatchevents/describe_rule.html} for full documentation.
}
\keyword{internal}
