% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevents_operations.R
\name{cloudwatchevents_update_connection}
\alias{cloudwatchevents_update_connection}
\title{Updates settings for a connection}
\usage{
cloudwatchevents_update_connection(
  Name,
  Description = NULL,
  AuthorizationType = NULL,
  AuthParameters = NULL
)
}
\arguments{
\item{Name}{[required] The name of the connection to update.}

\item{Description}{A description for the connection.}

\item{AuthorizationType}{The type of authorization to use for the connection.}

\item{AuthParameters}{The authorization parameters to use for the connection.}
}
\description{
Updates settings for a connection.

See \url{https://paws-r.github.io/docs/cloudwatchevents/update_connection.html} for full documentation.
}
\keyword{internal}
