% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_delete_feature}
\alias{cloudwatchevidently_delete_feature}
\title{Deletes an Evidently feature}
\usage{
cloudwatchevidently_delete_feature(feature, project)
}
\arguments{
\item{feature}{[required] The name of the feature to delete.}

\item{project}{[required] The name or ARN of the project that contains the feature to delete.}
}
\description{
Deletes an Evidently feature.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/delete_feature.html} for full documentation.
}
\keyword{internal}
