% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_delete_project}
\alias{cloudwatchevidently_delete_project}
\title{Deletes an Evidently project}
\usage{
cloudwatchevidently_delete_project(project)
}
\arguments{
\item{project}{[required] The name or ARN of the project to delete.}
}
\description{
Deletes an Evidently project. Before you can delete a project, you must delete all the features that the project contains. To delete a feature, use \code{\link[=cloudwatchevidently_delete_feature]{delete_feature}}.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/delete_project.html} for full documentation.
}
\keyword{internal}
