% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_get_experiment}
\alias{cloudwatchevidently_get_experiment}
\title{Returns the details about one experiment}
\usage{
cloudwatchevidently_get_experiment(experiment, project)
}
\arguments{
\item{experiment}{[required] The name of the experiment that you want to see the details of.}

\item{project}{[required] The name or ARN of the project that contains the experiment.}
}
\description{
Returns the details about one experiment. You must already know the experiment name. To retrieve a list of experiments in your account, use \code{\link[=cloudwatchevidently_list_experiments]{list_experiments}}.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/get_experiment.html} for full documentation.
}
\keyword{internal}
