% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchevidently_operations.R
\name{cloudwatchevidently_list_experiments}
\alias{cloudwatchevidently_list_experiments}
\title{Returns configuration details about all the experiments in the specified
project}
\usage{
cloudwatchevidently_list_experiments(
  maxResults = NULL,
  nextToken = NULL,
  project,
  status = NULL
)
}
\arguments{
\item{maxResults}{The maximum number of results to include in the response.}

\item{nextToken}{The token to use when requesting the next set of results. You received
this token from a previous
\code{\link[=cloudwatchevidently_list_experiments]{list_experiments}} operation.}

\item{project}{[required] The name or ARN of the project to return the experiment list from.}

\item{status}{Use this optional parameter to limit the returned results to only the
experiments with the status that you specify here.}
}
\description{
Returns configuration details about all the experiments in the specified project.

See \url{https://paws-r.github.io/docs/cloudwatchevidently/list_experiments.html} for full documentation.
}
\keyword{internal}
