% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_put_destination}
\alias{cloudwatchlogs_put_destination}
\title{Creates or updates a destination}
\usage{
cloudwatchlogs_put_destination(destinationName, targetArn, roleArn)
}
\arguments{
\item{destinationName}{[required] A name for the destination.}

\item{targetArn}{[required] The ARN of an Amazon Kinesis stream to which to deliver matching log
events.}

\item{roleArn}{[required] The ARN of an IAM role that grants CloudWatch Logs permissions to call
the Amazon Kinesis \code{PutRecord} operation on the destination stream.}
}
\description{
Creates or updates a destination. This operation is used only to create destinations for cross-account subscriptions.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/put_destination.html} for full documentation.
}
\keyword{internal}
