% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudwatchlogs_operations.R
\name{cloudwatchlogs_tag_log_group}
\alias{cloudwatchlogs_tag_log_group}
\title{Adds or updates the specified tags for the specified log group}
\usage{
cloudwatchlogs_tag_log_group(logGroupName, tags)
}
\arguments{
\item{logGroupName}{[required] The name of the log group.}

\item{tags}{[required] The key-value pairs to use for the tags.}
}
\description{
Adds or updates the specified tags for the specified log group.

See \url{https://paws-r.github.io/docs/cloudwatchlogs/tag_log_group.html} for full documentation.
}
\keyword{internal}
