% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_config_rule}
\alias{configservice_delete_config_rule}
\title{Deletes the specified Config rule and all of its evaluation results}
\usage{
configservice_delete_config_rule(ConfigRuleName)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule that you want to delete.}
}
\description{
Deletes the specified Config rule and all of its evaluation results.

See \url{https://paws-r.github.io/docs/configservice/delete_config_rule.html} for full documentation.
}
\keyword{internal}
