% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_delivery_channel}
\alias{configservice_delete_delivery_channel}
\title{Deletes the delivery channel}
\usage{
configservice_delete_delivery_channel(DeliveryChannelName)
}
\arguments{
\item{DeliveryChannelName}{[required] The name of the delivery channel to delete.}
}
\description{
Deletes the delivery channel.

See \url{https://paws-r.github.io/docs/configservice/delete_delivery_channel.html} for full documentation.
}
\keyword{internal}
