% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_evaluation_results}
\alias{configservice_delete_evaluation_results}
\title{Deletes the evaluation results for the specified Config rule}
\usage{
configservice_delete_evaluation_results(ConfigRuleName)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule for which you want to delete the evaluation
results.}
}
\description{
Deletes the evaluation results for the specified Config rule. You can specify one Config rule per request. After you delete the evaluation results, you can call the \code{\link[=configservice_start_config_rules_evaluation]{start_config_rules_evaluation}} API to start evaluating your Amazon Web Services resources against the rule.

See \url{https://paws-r.github.io/docs/configservice/delete_evaluation_results.html} for full documentation.
}
\keyword{internal}
