% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_organization_config_rule}
\alias{configservice_delete_organization_config_rule}
\title{Deletes the specified organization Config rule and all of its evaluation
results from all member accounts in that organization}
\usage{
configservice_delete_organization_config_rule(OrganizationConfigRuleName)
}
\arguments{
\item{OrganizationConfigRuleName}{[required] The name of organization Config rule that you want to delete.}
}
\description{
Deletes the specified organization Config rule and all of its evaluation results from all member accounts in that organization.

See \url{https://paws-r.github.io/docs/configservice/delete_organization_config_rule.html} for full documentation.
}
\keyword{internal}
