% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_organization_conformance_pack}
\alias{configservice_delete_organization_conformance_pack}
\title{Deletes the specified organization conformance pack and all of the
Config rules and remediation actions from all member accounts in that
organization}
\usage{
configservice_delete_organization_conformance_pack(
  OrganizationConformancePackName
)
}
\arguments{
\item{OrganizationConformancePackName}{[required] The name of organization conformance pack that you want to delete.}
}
\description{
Deletes the specified organization conformance pack and all of the Config rules and remediation actions from all member accounts in that organization.

See \url{https://paws-r.github.io/docs/configservice/delete_organization_conformance_pack.html} for full documentation.
}
\keyword{internal}
