% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_delete_remediation_exceptions}
\alias{configservice_delete_remediation_exceptions}
\title{Deletes one or more remediation exceptions mentioned in the resource
keys}
\usage{
configservice_delete_remediation_exceptions(ConfigRuleName, ResourceKeys)
}
\arguments{
\item{ConfigRuleName}{[required] The name of the Config rule for which you want to delete remediation
exception configuration.}

\item{ResourceKeys}{[required] An exception list of resource exception keys to be processed with the
current request. Config adds exception for each resource key. For
example, Config adds 3 exceptions for 3 resource keys.}
}
\description{
Deletes one or more remediation exceptions mentioned in the resource keys.

See \url{https://paws-r.github.io/docs/configservice/delete_remediation_exceptions.html} for full documentation.
}
\keyword{internal}
