% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_configuration_aggregators}
\alias{configservice_describe_configuration_aggregators}
\title{Returns the details of one or more configuration aggregators}
\usage{
configservice_describe_configuration_aggregators(
  ConfigurationAggregatorNames = NULL,
  NextToken = NULL,
  Limit = NULL
)
}
\arguments{
\item{ConfigurationAggregatorNames}{The name of the configuration aggregators.}

\item{NextToken}{The \code{nextToken} string returned on a previous page that you use to get
the next page of results in a paginated response.}

\item{Limit}{The maximum number of configuration aggregators returned on each page.
The default is maximum. If you specify 0, Config uses the default.}
}
\description{
Returns the details of one or more configuration aggregators. If the configuration aggregator is not specified, this action returns the details for all the configuration aggregators associated with the account.

See \url{https://paws-r.github.io/docs/configservice/describe_configuration_aggregators.html} for full documentation.
}
\keyword{internal}
