% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_configuration_recorders}
\alias{configservice_describe_configuration_recorders}
\title{Returns the details for the specified configuration recorders}
\usage{
configservice_describe_configuration_recorders(
  ConfigurationRecorderNames = NULL
)
}
\arguments{
\item{ConfigurationRecorderNames}{A list of configuration recorder names.}
}
\description{
Returns the details for the specified configuration recorders. If the configuration recorder is not specified, this action returns the details for all configuration recorders associated with the account.

See \url{https://paws-r.github.io/docs/configservice/describe_configuration_recorders.html} for full documentation.
}
\keyword{internal}
