% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_describe_conformance_pack_status}
\alias{configservice_describe_conformance_pack_status}
\title{Provides one or more conformance packs deployment status}
\usage{
configservice_describe_conformance_pack_status(
  ConformancePackNames = NULL,
  Limit = NULL,
  NextToken = NULL
)
}
\arguments{
\item{ConformancePackNames}{Comma-separated list of conformance pack names.}

\item{Limit}{The maximum number of conformance packs status returned on each page.}

\item{NextToken}{The \code{nextToken} string returned in a previous request that you use to
request the next page of results in a paginated response.}
}
\description{
Provides one or more conformance packs deployment status.

See \url{https://paws-r.github.io/docs/configservice/describe_conformance_pack_status.html} for full documentation.
}
\keyword{internal}
