% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_put_stored_query}
\alias{configservice_put_stored_query}
\title{Saves a new query or updates an existing saved query}
\usage{
configservice_put_stored_query(StoredQuery, Tags = NULL)
}
\arguments{
\item{StoredQuery}{[required] A list of \code{StoredQuery} objects. The mandatory fields are \code{QueryName}
and \code{Expression}.

When you are creating a query, you must provide a query name and an
expression. When you are updating a query, you must provide a query name
but updating the description is optional.}

\item{Tags}{A list of \code{Tags} object.}
}
\description{
Saves a new query or updates an existing saved query. The \code{QueryName} must be unique for a single Amazon Web Services account and a single Amazon Web Services Region. You can create upto 300 queries in a single Amazon Web Services account and a single Amazon Web Services Region.

See \url{https://paws-r.github.io/docs/configservice/put_stored_query.html} for full documentation.
}
\keyword{internal}
