% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configservice_operations.R
\name{configservice_untag_resource}
\alias{configservice_untag_resource}
\title{Deletes specified tags from a resource}
\usage{
configservice_untag_resource(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required] The Amazon Resource Name (ARN) that identifies the resource for which to
list the tags. Currently, the supported resources are \code{ConfigRule},
\code{ConfigurationAggregator} and \code{AggregatorAuthorization}.}

\item{TagKeys}{[required] The keys of the tags to be removed.}
}
\description{
Deletes specified tags from a resource.

See \url{https://paws-r.github.io/docs/configservice/untag_resource.html} for full documentation.
}
\keyword{internal}
