% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_tag_resource}
\alias{finspace_tag_resource}
\title{Adds metadata tags to a FinSpace resource}
\usage{
finspace_tag_resource(resourceArn, tags)
}
\arguments{
\item{resourceArn}{[required] The Amazon Resource Name (ARN) for the resource.}

\item{tags}{[required] One or more tags to be assigned to the resource.}
}
\description{
Adds metadata tags to a FinSpace resource.

See \url{https://paws-r.github.io/docs/finspace/tag_resource.html} for full documentation.
}
\keyword{internal}
