% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finspace_operations.R
\name{finspace_update_environment}
\alias{finspace_update_environment}
\title{Update your FinSpace environment}
\usage{
finspace_update_environment(
  environmentId,
  name = NULL,
  description = NULL,
  federationMode = NULL,
  federationParameters = NULL
)
}
\arguments{
\item{environmentId}{[required] The identifier of the FinSpace environment.}

\item{name}{The name of the environment.}

\item{description}{The description of the environment.}

\item{federationMode}{Authentication mode for the environment.
\itemize{
\item \code{FEDERATED} - Users access FinSpace through Single Sign On (SSO) via
your Identity provider.
\item \code{LOCAL} - Users access FinSpace via email and password managed
within the FinSpace environment.
}}

\item{federationParameters}{}
}
\description{
Update your FinSpace environment.

See \url{https://paws-r.github.io/docs/finspace/update_environment.html} for full documentation.
}
\keyword{internal}
