% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_operations.R
\name{health_disable_health_service_access_for_organization}
\alias{health_disable_health_service_access_for_organization}
\title{Disables Health from working with Organizations}
\usage{
health_disable_health_service_access_for_organization()
}
\description{
Disables Health from working with Organizations. To call this operation, you must sign in as an Identity and Access Management (IAM) user, assume an IAM role, or sign in as the root user (not recommended) in the organization's management account. For more information, see \href{https://docs.aws.amazon.com/health/latest/ug/aggregate-events.html}{Aggregating Health events} in the \emph{Health User Guide}.

See \url{https://paws-r.github.io/docs/health/disable_health_service_access_for_organization.html} for full documentation.
}
\keyword{internal}
