% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_check_in_license}
\alias{licensemanager_check_in_license}
\title{Checks in the specified license}
\usage{
licensemanager_check_in_license(LicenseConsumptionToken, Beneficiary = NULL)
}
\arguments{
\item{LicenseConsumptionToken}{[required] License consumption token.}

\item{Beneficiary}{License beneficiary.}
}
\description{
Checks in the specified license. Check in a license when it is no longer in use.

See \url{https://paws-r.github.io/docs/licensemanager/check_in_license.html} for full documentation.
}
\keyword{internal}
