% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_checkout_license}
\alias{licensemanager_checkout_license}
\title{Checks out the specified license}
\usage{
licensemanager_checkout_license(
  ProductSKU,
  CheckoutType,
  KeyFingerprint,
  Entitlements,
  ClientToken,
  Beneficiary = NULL,
  NodeId = NULL
)
}
\arguments{
\item{ProductSKU}{[required] Product SKU.}

\item{CheckoutType}{[required] Checkout type.}

\item{KeyFingerprint}{[required] Key fingerprint identifying the license.}

\item{Entitlements}{[required] License entitlements.}

\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{Beneficiary}{License beneficiary.}

\item{NodeId}{Node ID.}
}
\description{
Checks out the specified license.

See \url{https://paws-r.github.io/docs/licensemanager/checkout_license.html} for full documentation.
}
\keyword{internal}
