% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_create_grant_version}
\alias{licensemanager_create_grant_version}
\title{Creates a new version of the specified grant}
\usage{
licensemanager_create_grant_version(
  ClientToken,
  GrantArn,
  GrantName = NULL,
  AllowedOperations = NULL,
  Status = NULL,
  StatusReason = NULL,
  SourceVersion = NULL
)
}
\arguments{
\item{ClientToken}{[required] Unique, case-sensitive identifier that you provide to ensure the
idempotency of the request.}

\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}

\item{GrantName}{Grant name.}

\item{AllowedOperations}{Allowed operations for the grant.}

\item{Status}{Grant status.}

\item{StatusReason}{Grant status reason.}

\item{SourceVersion}{Current version of the grant.}
}
\description{
Creates a new version of the specified grant.

See \url{https://paws-r.github.io/docs/licensemanager/create_grant_version.html} for full documentation.
}
\keyword{internal}
