% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_grant}
\alias{licensemanager_delete_grant}
\title{Deletes the specified grant}
\usage{
licensemanager_delete_grant(GrantArn, StatusReason = NULL, Version)
}
\arguments{
\item{GrantArn}{[required] Amazon Resource Name (ARN) of the grant.}

\item{StatusReason}{The Status reason for the delete request.}

\item{Version}{[required] Current version of the grant.}
}
\description{
Deletes the specified grant.

See \url{https://paws-r.github.io/docs/licensemanager/delete_grant.html} for full documentation.
}
\keyword{internal}
