% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_delete_license_configuration}
\alias{licensemanager_delete_license_configuration}
\title{Deletes the specified license configuration}
\usage{
licensemanager_delete_license_configuration(LicenseConfigurationArn)
}
\arguments{
\item{LicenseConfigurationArn}{[required] ID of the license configuration.}
}
\description{
Deletes the specified license configuration.

See \url{https://paws-r.github.io/docs/licensemanager/delete_license_configuration.html} for full documentation.
}
\keyword{internal}
