% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license}
\alias{licensemanager_get_license}
\title{Gets detailed information about the specified license}
\usage{
licensemanager_get_license(LicenseArn, Version = NULL)
}
\arguments{
\item{LicenseArn}{[required] Amazon Resource Name (ARN) of the license.}

\item{Version}{License version.}
}
\description{
Gets detailed information about the specified license.

See \url{https://paws-r.github.io/docs/licensemanager/get_license.html} for full documentation.
}
\keyword{internal}
