% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_get_license_manager_report_generator}
\alias{licensemanager_get_license_manager_report_generator}
\title{Gets information about the specified report generator}
\usage{
licensemanager_get_license_manager_report_generator(
  LicenseManagerReportGeneratorArn
)
}
\arguments{
\item{LicenseManagerReportGeneratorArn}{[required] Amazon Resource Name (ARN) of the report generator.}
}
\description{
Gets information about the specified report generator.

See \url{https://paws-r.github.io/docs/licensemanager/get_license_manager_report_generator.html} for full documentation.
}
\keyword{internal}
