% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanager_operations.R
\name{licensemanager_list_received_grants}
\alias{licensemanager_list_received_grants}
\title{Lists grants that are received but not accepted}
\usage{
licensemanager_list_received_grants(
  GrantArns = NULL,
  Filters = NULL,
  NextToken = NULL,
  MaxResults = NULL
)
}
\arguments{
\item{GrantArns}{Amazon Resource Names (ARNs) of the grants.}

\item{Filters}{Filters to scope the results. The following filters are supported:
\itemize{
\item \code{ProductSKU}
\item \code{LicenseIssuerName}
\item \code{LicenseArn}
\item \code{GrantStatus}
\item \code{GranterAccountId}
}}

\item{NextToken}{Token for the next set of results.}

\item{MaxResults}{Maximum number of results to return in a single call.}
}
\description{
Lists grants that are received but not accepted.

See \url{https://paws-r.github.io/docs/licensemanager/list_received_grants.html} for full documentation.
}
\keyword{internal}
