% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_list_instances}
\alias{licensemanagerusersubscriptions_list_instances}
\title{Lists the EC2 instances providing user-based subscriptions}
\usage{
licensemanagerusersubscriptions_list_instances(
  Filters = NULL,
  MaxResults = NULL,
  NextToken = NULL
)
}
\arguments{
\item{Filters}{An array of structures that you can use to filter the results to those
that match one or more sets of key-value pairs that you specify.}

\item{MaxResults}{Maximum number of results to return in a single call.}

\item{NextToken}{Token for the next set of results.}
}
\description{
Lists the EC2 instances providing user-based subscriptions.

See \url{https://paws-r.github.io/docs/licensemanagerusersubscriptions/list_instances.html} for full documentation.
}
\keyword{internal}
