% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_register_identity_provider}
\alias{licensemanagerusersubscriptions_register_identity_provider}
\title{Registers an identity provider for user-based subscriptions}
\usage{
licensemanagerusersubscriptions_register_identity_provider(
  IdentityProvider,
  Product
)
}
\arguments{
\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{Product}{[required] The name of the user-based subscription product.}
}
\description{
Registers an identity provider for user-based subscriptions.

See \url{https://paws-r.github.io/docs/licensemanagerusersubscriptions/register_identity_provider.html} for full documentation.
}
\keyword{internal}
