% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/licensemanagerusersubscriptions_operations.R
\name{licensemanagerusersubscriptions_start_product_subscription}
\alias{licensemanagerusersubscriptions_start_product_subscription}
\title{Starts a product subscription for a user with the specified identity
provider}
\usage{
licensemanagerusersubscriptions_start_product_subscription(
  Domain = NULL,
  IdentityProvider,
  Product,
  Username
)
}
\arguments{
\item{Domain}{The domain name of the user.}

\item{IdentityProvider}{[required] An object that specifies details for the identity provider.}

\item{Product}{[required] The name of the user-based subscription product.}

\item{Username}{[required] The user name from the identity provider of the user.}
}
\description{
Starts a product subscription for a user with the specified identity provider.

See \url{https://paws-r.github.io/docs/licensemanagerusersubscriptions/start_product_subscription.html} for full documentation.
}
\keyword{internal}
