% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_associate_license}
\alias{managedgrafana_associate_license}
\title{Assigns a Grafana Enterprise license to a workspace}
\usage{
managedgrafana_associate_license(licenseType, workspaceId)
}
\arguments{
\item{licenseType}{[required] The type of license to associate with the workspace.}

\item{workspaceId}{[required] The ID of the workspace to associate the license with.}
}
\description{
Assigns a Grafana Enterprise license to a workspace. Upgrading to Grafana Enterprise incurs additional fees. For more information, see \href{https://docs.aws.amazon.com/grafana/latest/userguide/upgrade-to-Grafana-Enterprise.html}{Upgrade a workspace to Grafana Enterprise}.

See \url{https://paws-r.github.io/docs/managedgrafana/associate_license.html} for full documentation.
}
\keyword{internal}
