% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_describe_workspace}
\alias{managedgrafana_describe_workspace}
\title{Displays information about one Amazon Managed Grafana workspace}
\usage{
managedgrafana_describe_workspace(workspaceId)
}
\arguments{
\item{workspaceId}{[required] The ID of the workspace to display information about.}
}
\description{
Displays information about one Amazon Managed Grafana workspace.

See \url{https://paws-r.github.io/docs/managedgrafana/describe_workspace.html} for full documentation.
}
\keyword{internal}
