% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/managedgrafana_operations.R
\name{managedgrafana_update_permissions}
\alias{managedgrafana_update_permissions}
\title{Updates which users in a workspace have the Grafana Admin or Editor
roles}
\usage{
managedgrafana_update_permissions(updateInstructionBatch, workspaceId)
}
\arguments{
\item{updateInstructionBatch}{[required] An array of structures that contain the permission updates to make.}

\item{workspaceId}{[required] The ID of the workspace to update.}
}
\description{
Updates which users in a workspace have the Grafana \code{Admin} or \code{Editor} roles.

See \url{https://paws-r.github.io/docs/managedgrafana/update_permissions.html} for full documentation.
}
\keyword{internal}
